C10 - a Chef can Delete Recepie completed with Recepie_Ingredient table connection

+++ src/main/java/com/project/catering/api/RecepieEndpoint.java
@@ -75,7 +75,7 @@ public class RecepieEndpoint {
 	@DELETE
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.TEXT_PLAIN)
-	public void deleteRecepie(Long recepieid){
+	public void deleteRecepie(int recepieid){
 		recepieService.deleteRecepie(recepieid);
 	}
